requirements.txt
kaggle==1.5.12
numpy==1.23.4
pandas==1.5.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mpl-chord-diagram==0.4.0
pylint==2.15.5
black==22.10.0
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
db-dtypes==1.0.4
dataframe-image==0.1.3